Synchronise hour and minute hands of RadialTimePicker

+++ Code/RadialControls/Utilities/Conversion/MinutesDegreesConverter.cs
@@ -7,7 +7,7 @@ namespace RadialControls.Utilities
     {
         public object Convert(object value, Type targetType, object parameter, string language)
         {
-            return (((int) value) / 60) * 360;
+            return (((double) value) / 60) * 360;
         }
 
         public object ConvertBack(object value, Type targetType, object parameter, string language)
